refact: update console messagens in app.js
