+++ src/app.js
@@ -1,8 +1,10 @@
 const express = require('express');
 const { ApolloServer } = require('apollo-server-express');
+
 const typeDefs = require('./schemas/rootSchemas');
 const resolvers = require('./resolvers/rootResolvers');
 const models = require('../models');
+const testConnection = require('./util/testConnection');
 
 const server = new ApolloServer({
     typeDefs,
@@ -17,19 +19,6 @@ server.start()
         server.applyMiddleware({ app });
     });
 
-const testConnection = async () => {
-    try {
-        await models.sequelize.authenticate();
-        
-        console.log('Conexão estabelecida com sucesso.');
-        
-        await models.sequelize.sync();
-        
-        console.log('Todos os módulos foram sincronizados com sucesso.');
-    } catch (error) {
-        console.error('Não foi possível conectar ao banco de dados:', error.message);
-    }
-};
 testConnection();
 
 module.exports = app;
